Refactored out header screen
